+++ hashicorp-terraform/modules/helpers/TLS/providers.tf
@@ -1,0 +1,10 @@
+terraform {
+  experiments = [ module_variable_optional_attrs ]
+  
+  required_providers {
+    tls = {
+      source = "hashicorp/tls"
+      version = "~> 3.3"
+    }
+  }
+}